GUI Button
Added an on screen GUI Jump button for future Mobile builds.

Not working as intended. The GUI button only works if animations are
caused by a Trigger, which causes the jumps to happen in an odd manner.

Preferably, animations would be caused by a Boolean, but the GUI button
does not work when this is done.