Fixed On-Screen Jump Button
While I was unable to make the animations trigger using booleans, I
managed to create a regular UI Button which calls a new Jump function
OnClick.

The new Jump function contains if statements to trigger either the
regular or running jump. The regular jump requires that translation = 0,
while the running jump requires that translation != 0.

In this way, the new UI Button will only trigger the appropriate jump
and not both at the same time.