Drastic Changes
Because of how character movement was previously coded, it became
difficult to make the characters rigid body fall to the ground as
intended.

For this reason the character controller movement had to be completely
reworked. The previous version of the game was backed up to a separate
Unity project folder called Game Legacy.

I downloaded a larger set of animations for the character and used
several blend trees to control transitions from one to the other. Using
root motion I was able to move the character through the animations
themselves.

This method allowed me to ground the character during movement.

The HealthBar/Death Animation and the Joystick had to be modified to
cater for the changes in the movement controller.

Currently there is commented alternate code in the Character controller
for movement. The alternate code is intended for Joystick Movement as
opposed to Keyboard movement as it is difficult to have both working at
the same time.

Therefore you will have to comment/uncomment the appropriate code
depending on whether you want a PC or Mobile build of the game.

Default is set to Keyboard Movement.