Terrain
Created and painted a Terrain with a blend of Grass/Hill/Muddy Rock
textures.

Added several cubes which act as two bridges across a section of the
terrain, currently not textured.

The player model was scaled upwards, and the camera settings were
adjusted to compensate for the increased player size.