Main Building Addition
Created a model for a temple building in Maya 2016 and then exported to
Unity.

The building was scaled and placed into the scene. The model currently
has no textures.

I plan to create a few smaller buildings and then create textures for
the buildings, as well as paint some paved roads onto the terrain to
connect the buildings together.