Trash Models
Created four models in maya to be used as pick ups in the game:

Glass Bottle
Plastic Cup
Crumpled Paper
Metal box.

Unfortunately mental ray materials in Maya could not be exported to
Unity so I had to rely on simpler materials.

Also the crumpled paper was used by using nCloth on a plane and
animating it to crumple up like a piece of paper. However I was unable
to figure out how to successfully export this to Unity.  At the moment
it simply exports as a regular plane, without the crumpled up
appearance.

I am not sure I will be able to find time to fix this.