Menus
Created A Title Menu with its own terrain and UI.
Created an Instructions Menu with text explaining game objectives and
controls.
Added a Quit option to the Title Menu.

Created a ButtonManager script to control loading different scenes when
buttons are pressed.

Adjusted Canvas Scaler so that all UI is scaled based on screen size.

Organized Scenes in Build settings.