Joystick Fix
While going over the code, trying to eliminate any leftover lines and
scripts from previous builds, I realized that the joystick was no longer
working.

I adjusted the Virtual Joystick script so that it allowed for keyboard
input in case it did not detect any change in coordinates for the
joystick. This way I don't have to comment/uncomment code in the
Movement script whenever I want to switch between Keyboard/Joystick
Input.

As for the joystick issue, I spent a lot of time trying to figure out
the problem. It seemed that the inputVector value for the joystick
position float was not being assigned to the Animator Floats MoveZ and
Turn for some reason.

The issue was fixed after I unticked and then reticked "Apply Root
Motion" in the Animator component in the Inspector. I don't understand
why this was necessary, but it somehow fixed the issue.